fix: change zoom level input type from int to float when plotting.

+++ plot_run.py
@@ -7,12 +7,12 @@ parser = argparse.ArgumentParser(
     description='Generate hexagon heatmap from trajectory data')
 
 parser.add_argument('input_file', help='Input file path')
-parser.add_argument('-z', '--zoom', type=int, default=12,
+parser.add_argument('-z', '--zoom', type=float, default=12,
                     help='The zoom level of the map')
 args = parser.parse_args()
 
 
-#creat the plot directly from the hex sequence
+#create the plot directly from the hex sequence
 os.makedirs('out/plots', exist_ok=True)
 
 traj_seq = pd.read_csv(args.input_file)
